Refac: move PlayerActionHandler's logics to GameManager(which is facade) and separate ready logics(w. validation)
