+++ src/lib/client/game/index.ts
@@ -1,3 +1,4 @@
+import type { Block } from "$types";
 import type { EventBus } from "./event";
 import type { NetworkLayer } from "./network";
 import type { GameSequenceLayer } from "./sequence";
@@ -30,4 +31,32 @@ export class GameManager {
   private sequenceLayer: GameSequenceLayer;
   private networkLayer: NetworkLayer;
   private stateLayer: GameStateLayer;
+
+  submitMove({
+    previewUrl,
+    position,
+    blockInfo,
+    slotIdx,
+  }: {
+    previewUrl: string;
+    position: [number, number];
+    blockInfo: Block;
+    slotIdx: SlotIdx;
+  }) {
+    this.eventBus.publish('PlayerMoveSubmitted', {
+      previewUrl, position, blockInfo, slotIdx,
+    });
+  }
+
+  submitReady() {
+    this.eventBus.publish('PlayerReadySubmitted', undefined);
+  }
+
+  submitCancelReady() {
+    this.eventBus.publish('PlayerReadyCancelSubmitted', undefined);
+  }
+
+  startGame() {
+    this.eventBus.publish('GameStartRequested', undefined);
+  }
 }
